Esperanza Slashing (#184)

+++ src/esperanza/finalizationparams.cpp
@@ -32,7 +32,9 @@ bool ParseFinalizationParams(std::string jsonString,
     return false;
   }
 
-  paramsOut.m_epochLength = ParseNum(json, "epochLength", 50);
+  paramsOut.m_epochLength =
+      static_cast<uint32_t>(ParseNum(json, "epochLength", 50));
+
   paramsOut.m_minDepositSize = ParseNum(json, "minDepositSize", 1500 * UNIT);
   paramsOut.m_dynastyLogoutDelay = ParseNum(json, "dynastyLogoutDelay", 700);
   paramsOut.m_withdrawalEpochDelay =
